chore: get quote function and reorganize imports
