+++ packages/libs/client-utils/src/marmalade/buy-token.ts
@@ -13,16 +13,13 @@ import {
 import type { ChainId, IPactDecimal, IPactInt } from '@kadena/types';
 import { submitClient } from '../core';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type {
   CommonProps,
   IAuctionPurchaseConfig,
   IDutchAuctionPurchaseInput,
   WithAuctionPurchase,
 } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface IBuyTokenInput extends CommonProps {
   policyConfig?: { guarded: boolean };
