+++ packages/libs/client-utils/src/marmalade/create-auction.ts
@@ -9,10 +9,6 @@ import {
 import type { ChainId } from '@kadena/types';
 import { submitClient } from '../core';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type {
   CommonProps,
   IAuctionConfig,
@@ -20,6 +16,7 @@ import type {
   IDutchAuctionInput,
   WithAuction,
 } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface ICreateAuctionInput extends CommonProps {
   chainId: ChainId;
