+++ packages/libs/client-utils/src/marmalade/mint-token.ts
@@ -10,11 +10,8 @@ import {
 import type { ChainId, IPactDecimal } from '@kadena/types';
 import { submitClient } from '../core';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type { CommonProps } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface IMintTokenInput extends CommonProps {
   policyConfig?: {
