+++ packages/libs/client-utils/src/marmalade/place-bid.ts
@@ -11,16 +11,13 @@ import {
 import type { ChainId, IPactDecimal } from '@kadena/types';
 import { submitClient } from '../core';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type {
   CommonProps,
   IPlaceBidConfig,
   PlaceBidProps,
   WithPlaceBid,
 } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface IPlaceBidInput extends CommonProps {
   saleId: string;
