+++ packages/libs/client-utils/src/marmalade/transfer-token.ts
@@ -9,11 +9,8 @@ import {
 import type { ChainId, IPactDecimal } from '@kadena/types';
 import { submitClient } from '../core/client-helpers';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type { CommonProps } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface ITransferTokenInput extends CommonProps {
   policyConfig?: {
