+++ packages/libs/client-utils/src/marmalade/withdraw-token.ts
@@ -8,15 +8,12 @@ import {
 import type { ChainId, IPactDecimal, IPactInt } from '@kadena/types';
 import { submitClient } from '../core';
 import type { IClientConfig } from '../core/utils/helpers';
-import {
-  formatAdditionalSigners,
-  formatCapabilities,
-} from '../integration-tests/support/helpers';
 import type {
   CommonProps,
   IWithdrawSaleTokenPolicyConfig,
   WithWithdrawSaleTokenPolicy,
 } from './config';
+import { formatAdditionalSigners, formatCapabilities } from './helpers';
 
 interface IWithdrawTokenInput extends CommonProps {
   tokenId: string;
